+++ tests/test_annotations.py
@@ -103,7 +103,7 @@ class TestFlag:
         with pytest.warns(
             DeprecationWarning, match=re.escape("Flag.__init__'s `default` argument is deprecated, use Default instead")
         ):
-            flag = annotations.Flag(default=value)
+            flag = annotations.Flag(default=value)  # pyright: ignore [ reportDeprecated ]
 
         with pytest.warns(DeprecationWarning, match=re.escape("Use annotations.Default instead of the default arg")):
             assert flag.default is value
@@ -822,8 +822,8 @@ def test_with_generic_float_choices():
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        nom: typing.Annotated[annotations.Choices[Choices], "description"],
-        boom: typing.Annotated[annotations.Choices[Choices], "bag"] = Choices.Blam,
+        nom: typing.Annotated[annotations.Choices[Choices], "description"],  # type: ignore
+        boom: typing.Annotated[annotations.Choices[Choices], "bag"] = Choices.Blam,  # type: ignore
     ):
         ...
 
@@ -915,8 +915,8 @@ def test_with_generic_int_choices():
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        nat: typing.Annotated[annotations.Choices[Choices], "meow"],
-        bag: typing.Annotated[annotations.Choices[Choices], "bagette"] = Choices.Bazman,
+        nat: typing.Annotated[annotations.Choices[Choices], "meow"],  # type: ignore
+        bag: typing.Annotated[annotations.Choices[Choices], "bagette"] = Choices.Bazman,  # type: ignore
     ):
         ...
 
@@ -1007,8 +1007,8 @@ def test_with_generic_str_choices():
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        ny: typing.Annotated[annotations.Choices[Choices], "fat"],
-        aa: typing.Annotated[annotations.Choices[Choices], "bat"] = Choices.Sis,
+        ny: typing.Annotated[annotations.Choices[Choices], "fat"],  # type: ignore
+        aa: typing.Annotated[annotations.Choices[Choices], "bat"] = Choices.Sis,  # type: ignore
     ):
         ...
 
@@ -1189,8 +1189,8 @@ def test_with_generic_converted():
     @tanjun.as_message_command("nyaa")
     async def command(
         ctx: tanjun.abc.Context,
-        boo: typing.Annotated[annotations.Converted[mock_callback_1, mock_callback_2], "description"],  # type: ignore
-        bam: typing.Annotated[annotations.Converted[mock_callback_3], "nom"] = None,
+        boo: typing.Annotated[typing.Any, annotations.Converted(mock_callback_1, mock_callback_2), "description"],
+        bam: typing.Annotated[typing.Any, annotations.Converted(mock_callback_3), "nom"] = None,
     ) -> None:
         ...
 
@@ -1299,7 +1299,7 @@ def test_with_generic_default():
     @tanjun.as_message_command("name")
     async def command(
         ctx: tanjun.abc.Context,
-        argument: typing.Annotated[annotations.Default[annotations.Str, "nyaa"], "meow"],  # noqa: F821
+        argument: typing.Annotated[annotations.Default[annotations.Str, "nyaa"], "meow"],  # noqa: F821  # type: ignore
     ) -> None:
         ...
 
@@ -1342,7 +1342,7 @@ def test_with_default_overriding_signature_default():
     @tanjun.as_message_command("name")
     async def command(
         ctx: tanjun.abc.Context,
-        argument: typing.Annotated[annotations.Default[annotations.Str, "yeet"], "meow"] = "m",  # noqa: F821
+        argument: typing.Annotated[annotations.Str, annotations.Default("yeet"), "meow"] = "m",  # noqa: F821
     ) -> None:
         ...
 
@@ -1384,7 +1384,7 @@ def test_with_default_unsetting_signature_default():
     @tanjun.as_slash_command("name", "description")
     @tanjun.as_message_command("name")
     async def command(
-        ctx: tanjun.abc.Context, argument: typing.Annotated[annotations.Default[annotations.Str], "meow"] = "m"
+        ctx: tanjun.abc.Context, argument: typing.Annotated[annotations.Str, annotations.Default(), "meow"] = "m"
     ) -> None:
         ...
 
@@ -1487,7 +1487,9 @@ def test_with_flag_and_deprecated_default():
         @tanjun.as_slash_command("beep", "boop")
         async def callback(
             ctx: tanjun.abc.Context,
-            eep: typing.Annotated[annotations.Int, annotations.Flag(default=1231), "b"] = 545454,
+            eep: typing.Annotated[
+                annotations.Int, annotations.Flag(default=1231), "b"  # pyright: ignore [ reportDeprecated ]
+            ] = 545454,
         ) -> None:
             ...
 
@@ -1591,6 +1593,39 @@ def test_with_positional():
     assert option.default is tanjun.abc.NO_DEFAULT
 
 
+def test_with_generic_positional():
+    @annotations.with_annotated_args(follow_wrapped=True)
+    @tanjun.as_message_command("name")
+    @tanjun.as_slash_command("boop", "description")
+    async def callback(
+        ctx: tanjun.abc.Context, beep: typing.Annotated[annotations.Positional[annotations.Str], "eat"]  # type: ignore
+    ) -> None:
+        ...
+
+    assert isinstance(callback.parser, tanjun.ShlexParser)
+    assert isinstance(callback.wrapped_command, tanjun.SlashCommand)
+    assert len(callback.parser.arguments) == 1
+    assert len(callback.parser.options) == 0
+    option = callback.parser.arguments[0]
+    assert option.key == "beep"
+    assert option.converters == []
+    assert option.default is tanjun.abc.NO_DEFAULT
+    assert option.is_multi is False
+    assert option.min_length is None
+    assert option.max_length is None
+    assert option.min_value is None
+    assert option.max_value is None
+
+    assert callback.wrapped_command.build().options == [
+        hikari.CommandOption(
+            type=hikari.OptionType.STRING, name="beep", channel_types=None, description="eat", is_required=True
+        )
+    ]
+    assert len(callback.wrapped_command._tracked_options) == 1
+    option = callback.wrapped_command._tracked_options["beep"]
+    assert option.default is tanjun.abc.NO_DEFAULT
+
+
 def test_with_greedy():
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_message_command("command")
@@ -1614,7 +1649,7 @@ def test_with_greedy():
 def test_with_generic_greedy():
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_message_command("command")
-    async def callback(ctx: tanjun.abc.Context, meep: annotations.Greedy[annotations.Str]):
+    async def callback(ctx: tanjun.abc.Context, meep: annotations.Greedy[annotations.Str]):  # type: ignore
         ...
 
     assert isinstance(callback.parser, tanjun.ShlexParser)
@@ -1797,8 +1832,8 @@ def test_with_generic_length():
     @tanjun.as_message_command("name")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.Length[123], "nom"],
-        other_value: typing.Annotated[typing.Optional[annotations.Length[5544]], "meow"] = None,
+        value: typing.Annotated[annotations.Length[123], "nom"],  # type: ignore
+        other_value: typing.Annotated[typing.Optional[annotations.Length[5544]], "meow"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -1877,8 +1912,8 @@ def test_with_generic_length_when_min_specificed():
     @tanjun.as_message_command("name")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.Length[43, 5444], "nom"],
-        other_value: typing.Annotated[typing.Optional[annotations.Length[32, 4343]], "meow"] = None,
+        value: typing.Annotated[annotations.Length[43, 5444], "nom"],  # type: ignore
+        other_value: typing.Annotated[typing.Optional[annotations.Length[32, 4343]], "meow"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -2054,8 +2089,8 @@ def test_with_generic_max(
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        number: typing.Annotated[annotations.Max[value], "eee"],
-        other_number: typing.Annotated[annotations.Max[value], "eep"] = 54234,
+        number: typing.Annotated[annotations.Max[value], "eee"],  # type: ignore
+        other_number: typing.Annotated[annotations.Max[value], "eep"] = 54234,  # type: ignore
     ):
         ...
 
@@ -2321,8 +2356,8 @@ def test_with_generic_min(
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        number: typing.Annotated[annotations.Min[value], "bee"],
-        other_number: typing.Annotated[annotations.Min[value], "buzz"] = 321,
+        number: typing.Annotated[annotations.Min[value], "bee"],  # type: ignore
+        other_number: typing.Annotated[annotations.Min[value], "buzz"] = 321,  # type: ignore
     ):
         ...
 
@@ -2854,8 +2889,8 @@ def test_with_generic_ranged(
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        number: typing.Annotated[annotations.Ranged[min_value, max_value], "meow"],
-        other_number: typing.Annotated[annotations.Ranged[min_value, max_value], "nom"] = 443,
+        number: typing.Annotated[annotations.Ranged[min_value, max_value], "meow"],  # type: ignore
+        other_number: typing.Annotated[annotations.Ranged[min_value, max_value], "nom"] = 443,  # type: ignore
     ):
         ...
 
@@ -3015,8 +3050,8 @@ def test_with_generic_snowflake_or_for_channel():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Channel], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Channel]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Channel], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Channel]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3095,8 +3130,8 @@ def test_with_generic_snowflake_or_for_member():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Member], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Member]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Member], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Member]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3175,8 +3210,8 @@ def test_with_generic_snowflake_or_for_mentionable():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Mentionable], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Mentionable]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Mentionable], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Mentionable]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3255,8 +3290,8 @@ def test_with_generic_snowflake_or_for_role():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Role], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Role]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Role], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Role]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3335,8 +3370,8 @@ def test_with_generic_snowflake_or_for_user():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.User], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.User]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.User], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.User]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3415,8 +3450,8 @@ def test_with_generic_snowflake_or():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Bool], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Bool]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Bool], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Bool]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3615,9 +3650,9 @@ def test_with_generic_these_channels():
     @tanjun.as_message_command("name")
     async def command(
         ctx: tanjun.abc.Context,
-        bb: typing.Annotated[annotations.TheseChannels[hikari.GuildChannel], "nep"],
+        bb: typing.Annotated[annotations.TheseChannels[hikari.GuildChannel], "nep"],  # type: ignore
         bat: typing.Annotated[
-            typing.Optional[annotations.TheseChannels[hikari.GuildVoiceChannel, hikari.PrivateChannel]], "bip"
+            typing.Optional[annotations.TheseChannels[hikari.GuildVoiceChannel, hikari.PrivateChannel]], "bip"  # type: ignore
         ] = None,
     ):
         ...
@@ -4699,8 +4734,8 @@ def test_when_annotated_not_top_level():
     async def command(
         ctx: tanjun.abc.Context,
         *,
-        value: typing.Union[typing.Annotated[annotations.Positional[annotations.Str], "nyaa"], bool] = False,
-        other_value: typing.Optional[typing.Annotated[annotations.Ranged[123, 432], "meow"]] = None,
+        value: typing.Union[typing.Annotated[annotations.Str, annotations.Positional(), "nyaa"], bool] = False,
+        other_value: typing.Optional[typing.Annotated[annotations.Int, "meow"]] = None,
     ) -> None:
         raise NotImplementedError
 
@@ -4711,8 +4746,8 @@ def test_when_annotated_not_top_level():
             name="other_value",
             description="meow",
             is_required=False,
-            min_value=123,
-            max_value=432,
+            min_value=None,
+            max_value=None,
         ),
     ]
 
@@ -4760,8 +4795,8 @@ def test_when_annotated_not_top_level():
     assert option.is_multi is False
     assert option.min_length is None
     assert option.max_length is None
-    assert option.min_value == 123
-    assert option.max_value == 432
+    assert option.min_value is None
+    assert option.max_value is None
 
 
 if sys.version_info >= (3, 10):
@@ -4773,8 +4808,8 @@ if sys.version_info >= (3, 10):
         async def command(
             ctx: tanjun.abc.Context,
             *,
-            value: typing.Annotated[annotations.Positional[annotations.Str], "nyaa"] | bool = False,
-            other_value: typing.Annotated[annotations.Ranged[123, 432], "meow"] | None = None,
+            value: typing.Annotated[annotations.Str, annotations.Positional(), "nyaa"] | bool = False,
+            other_value: typing.Annotated[annotations.Int, "meow"] | None = None,
         ) -> None:
             raise NotImplementedError
 
@@ -4785,8 +4820,8 @@ if sys.version_info >= (3, 10):
                 name="other_value",
                 description="meow",
                 is_required=False,
-                min_value=123,
-                max_value=432,
+                min_value=None,
+                max_value=None,
             ),
         ]
 
@@ -4834,8 +4869,8 @@ if sys.version_info >= (3, 10):
         assert option.is_multi is False
         assert option.min_length is None
         assert option.max_length is None
-        assert option.min_value == 123
-        assert option.max_value == 432
+        assert option.min_value is None
+        assert option.max_value is None
 
 
 def test_when_annotated_handles_unions():
@@ -4845,8 +4880,8 @@ def test_when_annotated_handles_unions():
     async def command(
         ctx: tanjun.abc.Context,
         *,
-        value: typing.Annotated[typing.Union[annotations.Positional[annotations.Str], bool], "nyaa"] = False,
-        other_value: typing.Annotated[typing.Optional[annotations.Ranged[123, 432]], "meow"] = None,
+        value: typing.Annotated[typing.Union[annotations.Str, bool], annotations.Positional(), "nyaa"] = False,
+        other_value: typing.Annotated[typing.Optional[annotations.Int], "meow"] = None,
     ) -> None:
         raise NotImplementedError
 
@@ -4857,8 +4892,8 @@ def test_when_annotated_handles_unions():
             name="other_value",
             description="meow",
             is_required=False,
-            min_value=123,
-            max_value=432,
+            min_value=None,
+            max_value=None,
         ),
     ]
 
@@ -4906,8 +4941,8 @@ def test_when_annotated_handles_unions():
     assert option.is_multi is False
     assert option.min_length is None
     assert option.max_length is None
-    assert option.min_value == 123
-    assert option.max_value == 432
+    assert option.min_value is None
+    assert option.max_value is None
 
 
 if sys.version_info >= (3, 10):
@@ -4919,8 +4954,8 @@ if sys.version_info >= (3, 10):
         async def command(
             ctx: tanjun.abc.Context,
             *,
-            value: typing.Annotated[annotations.Positional[annotations.Str] | bool, "nyaa"] = False,
-            other_value: typing.Annotated[annotations.Ranged[123, 432] | None, "meow"] = None,
+            value: typing.Annotated[annotations.Str | bool, annotations.Positional(), "nyaa"] = False,
+            other_value: typing.Annotated[annotations.Int | None, "meow"] = None,
         ) -> None:
             raise NotImplementedError
 
@@ -4931,8 +4966,8 @@ if sys.version_info >= (3, 10):
                 name="other_value",
                 description="meow",
                 is_required=False,
-                min_value=123,
-                max_value=432,
+                min_value=None,
+                max_value=None,
             ),
         ]
 
@@ -4980,15 +5015,13 @@ if sys.version_info >= (3, 10):
         assert option.is_multi is False
         assert option.min_length is None
         assert option.max_length is None
-        assert option.min_value == 123
-        assert option.max_value == 432
+        assert option.min_value is None
+        assert option.max_value is None
 
 
 def test_parse_annotated_args_with_descriptions_argument():
     @tanjun.as_slash_command("name", "description")
-    async def command(
-        ctx: tanjun.abc.Context, *, echo: annotations.Str, foxy: annotations.Ranged[123, 432] = 232
-    ) -> None:
+    async def command(ctx: tanjun.abc.Context, *, echo: annotations.Str, foxy: annotations.Int = 232) -> None:
         raise NotImplementedError
 
     annotations.parse_annotated_args(command, descriptions={"echo": "meow", "foxy": "x3", "unknown": "..."})
@@ -5000,8 +5033,8 @@ def test_parse_annotated_args_with_descriptions_argument():
             name="foxy",
             description="x3",
             is_required=False,
-            min_value=123,
-            max_value=432,
+            min_value=None,
+            max_value=None,
         ),
     ]
 
@@ -5704,8 +5737,8 @@ def test_with_unpacked_typed_dict_and_converted():
     mock_callback_2 = mock.Mock()
 
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Converted[mock_callback_1], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Converted[mock_callback_2], "xat"]]
+        of: typing.Annotated[typing.Any, annotations.Converted(mock_callback_1), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[typing.Any, annotations.Converted(mock_callback_2), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -5832,8 +5865,8 @@ def test_with_unpacked_typed_dict_and_datetime():
 
 def test_with_unpacked_typed_dict_and_default():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Default[annotations.Int, 0], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Default[annotations.Float, 0.1], "xat"]]
+        of: typing.Annotated[annotations.Int, annotations.Default(0), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Float, annotations.Default(0.1), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6005,7 +6038,7 @@ def test_with_unpacked_typed_dict_and_float():
 
 def test_with_unpacked_typed_dict_and_greedy():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Greedy[annotations.Str], "maaaa"]
+        of: typing.Annotated[annotations.Str, annotations.Greedy(), "maaaa"]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6181,8 +6214,8 @@ def test_with_unpacked_typed_dict_and_interaction_member():
 
 def test_with_unpacked_typed_dict_and_length():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Length[232], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Length[4, 128], "xat"]]
+        of: typing.Annotated[annotations.Str, annotations.Length(232), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Str, annotations.Length(4, 128), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6254,8 +6287,8 @@ def test_with_unpacked_typed_dict_and_length():
 
 def test_with_unpacked_typed_dict_and_max():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Max[453], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Max[69.420], "xat"]]
+        of: typing.Annotated[annotations.Int, annotations.Max(453), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Float, annotations.Max(69.420), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6450,8 +6483,8 @@ def test_with_unpacked_typed_dict_and_mentionable():
 
 def test_with_unpacked_typed_dict_and_min():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Min[3.2], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Min[32], "xat"]]
+        of: typing.Annotated[annotations.Float, annotations.Min(3.2), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Int, annotations.Min(32), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6582,8 +6615,8 @@ def test_with_unpacked_typed_dict_and_name():
 
 def test_with_unpacked_typed_dict_and_positional():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Positional[annotations.Int], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Positional[annotations.Str], "xat"]]
+        of: typing.Annotated[annotations.Int, annotations.Positional(), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Str, annotations.Positional(), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_message_command("x", "3")
@@ -6620,8 +6653,8 @@ def test_with_unpacked_typed_dict_and_positional():
 
 def test_with_unpacked_typed_dict_and_ranged():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.Ranged[4, 64], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Ranged[12.21, 54.34], "xat"]]
+        of: typing.Annotated[annotations.Int, annotations.Ranged(4, 64), "maaaa"]
+        oo: typing_extensions.NotRequired[typing.Annotated[annotations.Float, annotations.Ranged(12.21, 54.34), "xat"]]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6821,8 +6854,18 @@ def test_with_unpacked_typed_dict_and_snowflake():
 
 def test_with_unpacked_typed_dict_and_snowflake_or():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.SnowflakeOr[annotations.Role], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.SnowflakeOr[annotations.Member], "xat"]]
+        of: typing.Annotated[
+            typing.Union[annotations.Role, hikari.Snowflake],
+            annotations.SnowflakeOr(parse_id=tanjun.conversion.parse_role_id),
+            "maaaa",
+        ]
+        oo: typing_extensions.NotRequired[
+            typing.Annotated[
+                typing.Union[annotations.Member, hikari.Snowflake],
+                annotations.SnowflakeOr(parse_id=tanjun.conversion.parse_user_id),
+                "xat",
+            ]
+        ]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")
@@ -6949,8 +6992,12 @@ def test_with_unpacked_typed_dict_and_str():
 
 def test_with_unpacked_typed_dict_and_these_channels():
     class TypedDict(typing_extensions.TypedDict):
-        of: typing.Annotated[annotations.TheseChannels[hikari.ChannelType.DM, hikari.GuildThreadChannel], "maaaa"]
-        oo: typing_extensions.NotRequired[typing.Annotated[annotations.TheseChannels[hikari.GuildTextChannel], "xat"]]
+        of: typing.Annotated[
+            annotations.Channel, annotations.TheseChannels(hikari.ChannelType.DM, hikari.GuildThreadChannel), "maaaa"
+        ]
+        oo: typing_extensions.NotRequired[
+            typing.Annotated[annotations.Channel, annotations.TheseChannels(hikari.GuildTextChannel), "xat"]
+        ]
 
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_slash_command("a", "b")