+++ tests/test_annotations_future_annotations.py
@@ -696,8 +696,8 @@ def test_with_generic_float_choices():
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        nom: typing.Annotated[annotations.Choices[Choices1], "description"],
-        boom: typing.Annotated[annotations.Choices[Choices1], "bag"] = Choices1.Blam,
+        nom: typing.Annotated[annotations.Choices[Choices1], "description"],  # type: ignore
+        boom: typing.Annotated[annotations.Choices[Choices1], "bag"] = Choices1.Blam,  # type: ignore
     ):
         ...
 
@@ -791,8 +791,8 @@ def test_with_generic_int_choices():
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        nat: typing.Annotated[annotations.Choices[Choices2], "meow"],
-        bag: typing.Annotated[annotations.Choices[Choices2], "bagette"] = Choices2.Bazman,
+        nat: typing.Annotated[annotations.Choices[Choices2], "meow"],  # type: ignore
+        bag: typing.Annotated[annotations.Choices[Choices2], "bagette"] = Choices2.Bazman,  # type: ignore
     ):
         ...
 
@@ -885,8 +885,8 @@ def test_with_generic_str_choices():
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        ny: typing.Annotated[annotations.Choices[Choices3], "fat"],
-        aa: typing.Annotated[annotations.Choices[Choices3], "bat"] = Choices3.Sis,
+        ny: typing.Annotated[annotations.Choices[Choices3], "fat"],  # type: ignore
+        aa: typing.Annotated[annotations.Choices[Choices3], "bat"] = Choices3.Sis,  # type: ignore
     ):
         ...
 
@@ -1074,7 +1074,7 @@ def test_with_generic_converted():
     async def command(
         ctx: tanjun.abc.Context,
         boo: typing.Annotated[annotations.Converted[mock_callback_1, mock_callback_2], "description"],  # type: ignore
-        bam: typing.Annotated[annotations.Converted[mock_callback_3], "nom"] = None,
+        bam: typing.Annotated[annotations.Converted[mock_callback_3], "nom"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -1183,7 +1183,7 @@ def test_with_generic_default():
     @tanjun.as_message_command("name")
     async def command(
         ctx: tanjun.abc.Context,
-        argument: typing.Annotated[annotations.Default[annotations.Str, "nyaa"], "meow"],  # noqa: F821
+        argument: typing.Annotated[annotations.Default[annotations.Str, "nyaa"], "meow"],  # noqa: F821  #type: ignore
     ) -> None:
         ...
 
@@ -1226,7 +1226,7 @@ def test_with_default_overriding_signature_default():
     @tanjun.as_message_command("name")
     async def command(
         ctx: tanjun.abc.Context,
-        argument: typing.Annotated[annotations.Default[annotations.Str, "yeet"], "meow"] = "m",  # noqa: F821
+        argument: typing.Annotated[annotations.Str, annotations.Default("yeet"), "meow"] = "m",  # noqa: F821
     ) -> None:
         ...
 
@@ -1268,7 +1268,7 @@ def test_with_default_unsetting_signature_default():
     @tanjun.as_slash_command("name", "description")
     @tanjun.as_message_command("name")
     async def command(
-        ctx: tanjun.abc.Context, argument: typing.Annotated[annotations.Default[annotations.Str], "meow"] = "m"
+        ctx: tanjun.abc.Context, argument: typing.Annotated[annotations.Str, annotations.Default(), "meow"] = "m"
     ) -> None:
         ...
 
@@ -1372,7 +1372,9 @@ def test_with_flag_and_deprecated_default():
         @tanjun.as_slash_command("beep", "boop")
         async def callback(
             ctx: tanjun.abc.Context,
-            eep: typing.Annotated[annotations.Int, annotations.Flag(default=1231), "b"] = 545454,
+            eep: typing.Annotated[
+                annotations.Int, annotations.Flag(default=1231), "b"  # pyright: ignore [ reportDeprecated ]
+            ] = 545454,
         ) -> None:
             ...
 
@@ -1476,6 +1478,39 @@ def test_with_positional():
     assert option.default is tanjun.abc.NO_DEFAULT
 
 
+def test_with_generic_positional():
+    @annotations.with_annotated_args(follow_wrapped=True)
+    @tanjun.as_message_command("name")
+    @tanjun.as_slash_command("boop", "description")
+    async def callback(
+        ctx: tanjun.abc.Context, beep: typing.Annotated[annotations.Positional[annotations.Str], "eat"]  # type: ignore
+    ) -> None:
+        ...
+
+    assert isinstance(callback.parser, tanjun.ShlexParser)
+    assert isinstance(callback.wrapped_command, tanjun.SlashCommand)
+    assert len(callback.parser.arguments) == 1
+    assert len(callback.parser.options) == 0
+    option = callback.parser.arguments[0]
+    assert option.key == "beep"
+    assert option.converters == []
+    assert option.default is tanjun.abc.NO_DEFAULT
+    assert option.is_multi is False
+    assert option.min_length is None
+    assert option.max_length is None
+    assert option.min_value is None
+    assert option.max_value is None
+
+    assert callback.wrapped_command.build().options == [
+        hikari.CommandOption(
+            type=hikari.OptionType.STRING, name="beep", channel_types=None, description="eat", is_required=True
+        )
+    ]
+    assert len(callback.wrapped_command._tracked_options) == 1
+    option = callback.wrapped_command._tracked_options["beep"]
+    assert option.default is tanjun.abc.NO_DEFAULT
+
+
 def test_with_greedy():
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_message_command("command")
@@ -1499,7 +1534,7 @@ def test_with_greedy():
 def test_with_generic_greedy():
     @annotations.with_annotated_args(follow_wrapped=True)
     @tanjun.as_message_command("command")
-    async def callback(ctx: tanjun.abc.Context, meep: annotations.Greedy[annotations.Str]):
+    async def callback(ctx: tanjun.abc.Context, meep: annotations.Greedy[annotations.Str]):  # type: ignore
         ...
 
     assert isinstance(callback.parser, tanjun.ShlexParser)
@@ -1682,8 +1717,8 @@ def test_with_generic_length():
     @tanjun.as_message_command("name")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.Length[123], "nom"],
-        other_value: typing.Annotated[typing.Optional[annotations.Length[5544]], "meow"] = None,
+        value: typing.Annotated[annotations.Length[123], "nom"],  # type: ignore
+        other_value: typing.Annotated[typing.Optional[annotations.Length[5544]], "meow"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -1762,8 +1797,8 @@ def test_with_generic_length_when_min_specificed():
     @tanjun.as_message_command("name")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.Length[43, 5444], "nom"],
-        other_value: typing.Annotated[typing.Optional[annotations.Length[32, 4343]], "meow"] = None,
+        value: typing.Annotated[annotations.Length[43, 5444], "nom"],  # type: ignore
+        other_value: typing.Annotated[typing.Optional[annotations.Length[32, 4343]], "meow"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -1947,8 +1982,8 @@ def test_with_generic_max(
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        number: typing.Annotated[annotations.Max[value_], "eee"],
-        other_number: typing.Annotated[annotations.Max[value_], "eep"] = 54234,
+        number: typing.Annotated[annotations.Max[value_], "eee"],  # type: ignore
+        other_number: typing.Annotated[annotations.Max[value_], "eep"] = 54234,  # type: ignore
     ):
         ...
 
@@ -2222,8 +2257,8 @@ def test_with_generic_min(
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        number: typing.Annotated[annotations.Min[value_], "bee"],
-        other_number: typing.Annotated[annotations.Min[value_], "buzz"] = 321,
+        number: typing.Annotated[annotations.Min[value_], "bee"],  # type: ignore
+        other_number: typing.Annotated[annotations.Min[value_], "buzz"] = 321,  # type: ignore
     ):
         ...
 
@@ -2760,8 +2795,8 @@ def test_with_generic_ranged(
     @tanjun.as_message_command("command")
     async def callback(
         ctx: tanjun.abc.Context,
-        number: typing.Annotated[annotations.Ranged[min_value_, max_value_], "meow"],
-        other_number: typing.Annotated[annotations.Ranged[min_value_, max_value_], "nom"] = 443,
+        number: typing.Annotated[annotations.Ranged[min_value_, max_value_], "meow"],  # type: ignore
+        other_number: typing.Annotated[annotations.Ranged[min_value_, max_value_], "nom"] = 443,  # type: ignore
     ):
         ...
 
@@ -2922,8 +2957,8 @@ def test_with_generic_snowflake_or_for_channel():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Channel], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Channel]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Channel], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Channel]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3002,8 +3037,8 @@ def test_with_generic_snowflake_or_for_member():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Member], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Member]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Member], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Member]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3082,8 +3117,8 @@ def test_with_generic_snowflake_or_for_mentionable():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Mentionable], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Mentionable]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Mentionable], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Mentionable]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3162,8 +3197,8 @@ def test_with_generic_snowflake_or_for_role():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Role], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Role]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Role], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Role]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3242,8 +3277,8 @@ def test_with_generic_snowflake_or_for_user():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.User], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.User]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.User], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.User]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3322,8 +3357,8 @@ def test_with_generic_snowflake_or():
     @tanjun.as_slash_command("yeet", "description")
     async def callback(
         ctx: tanjun.abc.Context,
-        value: typing.Annotated[annotations.SnowflakeOr[annotations.Bool], "se"],
-        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Bool]], "x"] = None,
+        value: typing.Annotated[annotations.SnowflakeOr[annotations.Bool], "se"],  # type: ignore
+        value_2: typing.Annotated[annotations.SnowflakeOr[typing.Optional[annotations.Bool]], "x"] = None,  # type: ignore
     ) -> None:
         ...
 
@@ -3525,9 +3560,9 @@ def test_with_generic_these_channels():
     @tanjun.as_message_command("name")
     async def command(
         ctx: tanjun.abc.Context,
-        bb: typing.Annotated[annotations.TheseChannels[hikari.GuildChannel], "nep"],
+        bb: typing.Annotated[annotations.TheseChannels[hikari.GuildChannel], "nep"],  # type: ignore
         bat: typing.Annotated[
-            typing.Optional[annotations.TheseChannels[hikari.GuildVoiceChannel, hikari.PrivateChannel]], "bip"
+            typing.Optional[annotations.TheseChannels[hikari.GuildVoiceChannel, hikari.PrivateChannel]], "bip"  # type: ignore
         ] = None,
     ):
         ...
@@ -4610,8 +4645,8 @@ def test_when_annotated_not_top_level():
     async def command(
         ctx: tanjun.abc.Context,
         *,
-        value: typing.Union[typing.Annotated[annotations.Positional[annotations.Str], "nyaa"], bool] = False,
-        other_value: typing.Optional[typing.Annotated[annotations.Ranged[123, 432], "meow"]] = None,
+        value: typing.Union[typing.Annotated[annotations.Str, annotations.Positional(), "nyaa"], bool] = False,
+        other_value: typing.Optional[typing.Annotated[annotations.Int, "meow"]] = None,
     ) -> None:
         raise NotImplementedError
 
@@ -4622,8 +4657,8 @@ def test_when_annotated_not_top_level():
             name="other_value",
             description="meow",
             is_required=False,
-            min_value=123,
-            max_value=432,
+            min_value=None,
+            max_value=None,
         ),
     ]
 
@@ -4671,8 +4706,8 @@ def test_when_annotated_not_top_level():
     assert option.is_multi is False
     assert option.min_length is None
     assert option.max_length is None
-    assert option.min_value == 123
-    assert option.max_value == 432
+    assert option.min_value is None
+    assert option.max_value is None
 
 
 if sys.version_info >= (3, 10):
@@ -4684,8 +4719,8 @@ if sys.version_info >= (3, 10):
         async def command(
             ctx: tanjun.abc.Context,
             *,
-            value: typing.Annotated[annotations.Positional[annotations.Str], "nyaa"] | bool = False,
-            other_value: typing.Annotated[annotations.Ranged[123, 432], "meow"] | None = None,
+            value: typing.Annotated[annotations.Str, annotations.Positional(), "nyaa"] | bool = False,
+            other_value: typing.Annotated[annotations.Int, "meow"] | None = None,
         ) -> None:
             raise NotImplementedError
 
@@ -4696,8 +4731,8 @@ if sys.version_info >= (3, 10):
                 name="other_value",
                 description="meow",
                 is_required=False,
-                min_value=123,
-                max_value=432,
+                min_value=None,
+                max_value=None,
             ),
         ]
 
@@ -4745,8 +4780,8 @@ if sys.version_info >= (3, 10):
         assert option.is_multi is False
         assert option.min_length is None
         assert option.max_length is None
-        assert option.min_value == 123
-        assert option.max_value == 432
+        assert option.min_value is None
+        assert option.max_value is None
 
 
 def test_when_annotated_handles_unions():
@@ -4756,8 +4791,8 @@ def test_when_annotated_handles_unions():
     async def command(
         ctx: tanjun.abc.Context,
         *,
-        value: typing.Annotated[typing.Union[annotations.Positional[annotations.Str], bool], "nyaa"] = False,
-        other_value: typing.Annotated[typing.Optional[annotations.Ranged[123, 432]], "meow"] = None,
+        value: typing.Annotated[typing.Union[annotations.Str, bool], annotations.Positional(), "nyaa"] = False,
+        other_value: typing.Annotated[typing.Optional[annotations.Int], "meow"] = None,
     ) -> None:
         raise NotImplementedError
 
@@ -4768,8 +4803,8 @@ def test_when_annotated_handles_unions():
             name="other_value",
             description="meow",
             is_required=False,
-            min_value=123,
-            max_value=432,
+            min_value=None,
+            max_value=None,
         ),
     ]
 
@@ -4817,8 +4852,8 @@ def test_when_annotated_handles_unions():
     assert option.is_multi is False
     assert option.min_length is None
     assert option.max_length is None
-    assert option.min_value == 123
-    assert option.max_value == 432
+    assert option.min_value is None
+    assert option.max_value is None
 
 
 if sys.version_info >= (3, 10):
@@ -4830,8 +4865,8 @@ if sys.version_info >= (3, 10):
         async def command(
             ctx: tanjun.abc.Context,
             *,
-            value: typing.Annotated[annotations.Positional[annotations.Str] | bool, "nyaa"] = False,
-            other_value: typing.Annotated[annotations.Ranged[123, 432] | None, "meow"] = None,
+            value: typing.Annotated[annotations.Str | bool, annotations.Positional(), "nyaa"] = False,
+            other_value: typing.Annotated[annotations.Int | None, "meow"] = None,
         ) -> None:
             raise NotImplementedError
 
@@ -4842,8 +4877,8 @@ if sys.version_info >= (3, 10):
                 name="other_value",
                 description="meow",
                 is_required=False,
-                min_value=123,
-                max_value=432,
+                min_value=None,
+                max_value=None,
             ),
         ]
 
@@ -4891,15 +4926,13 @@ if sys.version_info >= (3, 10):
         assert option.is_multi is False
         assert option.min_length is None
         assert option.max_length is None
-        assert option.min_value == 123
-        assert option.max_value == 432
+        assert option.min_value is None
+        assert option.max_value is None
 
 
 def test_parse_annotated_args_with_descriptions_argument():
     @tanjun.as_slash_command("name", "description")
-    async def command(
-        ctx: tanjun.abc.Context, *, echo: annotations.Str, foxy: annotations.Ranged[123, 432] = 232
-    ) -> None:
+    async def command(ctx: tanjun.abc.Context, *, echo: annotations.Str, foxy: annotations.Int = 232) -> None:
         raise NotImplementedError
 
     annotations.parse_annotated_args(command, descriptions={"echo": "meow", "foxy": "x3", "unknown": "..."})
@@ -4911,8 +4944,8 @@ def test_parse_annotated_args_with_descriptions_argument():
             name="foxy",
             description="x3",
             is_required=False,
-            min_value=123,
-            max_value=432,
+            min_value=None,
+            max_value=None,
         ),
     ]