+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Deprecated
+- Using the following classes as generic types (i.e. `Class[...]`):
+  * [yuyo.annotations.Converted][]
+  * [yuyo.annotations.Default][]
+  * [yuyo.annotations.Positional][]
+  * [yuyo.annotations.Greedy][]
+  * [yuyo.annotations.Length][]
+  * [yuyo.annotations.Max][]
+  * [yuyo.annotations.Min][]
+  * [yuyo.annotations.Ranged][]
+  * [yuyo.annotations.SnowflakeOr][]
+
 ## [2.13.0] - 2023-04-10
 ### Added
 - Support for sending stickers to just [MessageContext.respond][tanjun.abc.MessageContext.respond].